+++ history.md
@@ -43,6 +43,8 @@ Semantic Versioning 2.0.0 is from version 0.1.6+
 
 ## version 0.7.7 - _(Unreleased)_ - 2026-?-? {#v0.7.7}
 
+- [x] (Fixed) _Back-end_ Dropbox client wrapper test when running offline (PR #2603)
+- [x] (Fixed) _Back-end_ Add support for macOS and external drive in tests (PR #2602)
 - [x] (Fixed) _Back-end_ Sync Race Condition Fix (PR #2550)
 - [x] (Changed) _Back-end_ Upgrade to .NET 8 - SDK 8.0.417 (Runtime: 8.0.23) (PR #2588)
 - [x] (Added) _Front-end_ Batch rename feature (PR #2585) (Issue #72)